ref: create `connection` and `known_network` modules
