+++ crawler/src/main.rs
@@ -216,198 +216,3 @@ impl Crawler {
 }
 
 // == CONNECTIONS, todo: restructure crate.
-
-use parking_lot::RwLock;
-use std::{
-    cmp::Ordering,
-    collections::{HashMap, HashSet},
-    hash::{Hash, Hasher},
-};
-use time::OffsetDateTime;
-
-// Purges connections that haven't been seen within this time (in hours).
-const STALE_CONNECTION_CUTOFF_TIME_HRS: i64 = 4;
-
-/// A connection between two peers.
-///
-/// Implements `partialEq` and `Hash` manually so that the `source`-`target` order has no impact on equality
-/// (since connections are directionless). The timestamp is also not included in the comparison.
-#[derive(Debug, Eq, Copy, Clone)]
-pub struct Connection {
-    /// One side of the connection.
-    pub source: SocketAddr,
-    /// The other side of the connection.
-    pub target: SocketAddr,
-    /// The last time this peer was seen by the crawler (used determine which connections are
-    /// likely stale).
-    last_seen: OffsetDateTime,
-}
-
-impl PartialEq for Connection {
-    fn eq(&self, other: &Self) -> bool {
-        let (a, b) = (self.source, self.target);
-        let (c, d) = (other.source, other.target);
-
-        a == d && b == c || a == c && b == d
-    }
-}
-
-impl Hash for Connection {
-    fn hash<H: Hasher>(&self, state: &mut H) {
-        let (a, b) = (self.source, self.target);
-
-        // This ensures the hash is the same for (a, b) as it is for (b, a).
-        match a.cmp(&b) {
-            Ordering::Greater => {
-                b.hash(state);
-                a.hash(state);
-            }
-            _ => {
-                a.hash(state);
-                b.hash(state);
-            }
-        }
-    }
-}
-
-impl Connection {
-    pub fn new(source: SocketAddr, target: SocketAddr) -> Self {
-        Connection {
-            source,
-            target,
-            last_seen: OffsetDateTime::now_utc(),
-        }
-    }
-}
-
-/// Constructs a set of nodes contained from the connection set.
-pub fn nodes_from_connections(connections: &HashSet<Connection>) -> HashSet<SocketAddr> {
-    let mut nodes: HashSet<SocketAddr> = HashSet::new();
-    for connection in connections.iter() {
-        // Using a hashset guarantees uniqueness.
-        nodes.insert(connection.source);
-        nodes.insert(connection.target);
-    }
-
-    nodes
-}
-
-/// Keeps track of crawled peers and their connections.
-#[derive(Debug)]
-pub struct KnownNetwork {
-    // The nodes and their block height if known.
-    nodes: RwLock<HashMap<SocketAddr, u32>>,
-    // The connections map.
-    connections: RwLock<HashSet<Connection>>,
-}
-
-impl KnownNetwork {
-    // More convenient for testing.
-    fn update_connections(&self, source: SocketAddr, peers: Vec<SocketAddr>) {
-        // Rules:
-        //  - if a connecton exists already, do nothing.
-        //  - if a connection is new, add it.
-        //  - if an exisitng connection involving the source isn't in the peerlist, remove it if
-        //  it's stale.
-
-        let new_connections: HashSet<Connection> = peers.into_iter().map(|peer| Connection::new(source, peer)).collect();
-
-        // Find which connections need to be removed.
-        //
-        // With sets: a - b = removed connections (if and only if one of the two addrs is the
-        // source), otherwise it's a connection which doesn't include the source and shouldn't be
-        // removed. We also keep connections seen within the last few hours as peerlists are capped
-        // in size and omitted connections don't necessarily mean they don't exist anymore.
-        let connections_to_remove: HashSet<Connection> = self
-            .connections
-            .read()
-            .difference(&new_connections)
-            .filter(|conn| {
-                (conn.source == source || conn.target == source)
-                    && (OffsetDateTime::now_utc() - conn.last_seen).whole_hours() > STALE_CONNECTION_CUTOFF_TIME_HRS
-            })
-            .copied()
-            .collect();
-
-        // Scope the write lock.
-        {
-            let mut connections_g = self.connections.write();
-
-            // Remove stale connections.
-            connections_g.retain(|connection| !connections_to_remove.contains(connection));
-
-            // Insert new connections, we use replace so the last seen timestamp is overwritten.
-            for new_connection in new_connections.into_iter() {
-                connections_g.replace(new_connection);
-            }
-        }
-
-        // Scope the write lock.
-        {
-            let mut nodes_g = self.nodes.write();
-
-            // Remove the nodes that no longer correspond to connections.
-            let nodes_from_connections = nodes_from_connections(&self.connections());
-            nodes_g.retain(|addr, _| nodes_from_connections.contains(addr));
-        }
-    }
-
-    /// Update the height stored for this particular node.
-    pub fn update_height(&self, source: SocketAddr, height: u32) {
-        self.nodes.write().insert(source, height);
-    }
-
-    /// Returns `true` if the known network contains any connections, `false` otherwise.
-    pub fn has_connections(&self) -> bool {
-        !self.connections.read().is_empty()
-    }
-
-    /// Returns a connection.
-    pub fn get_connection(&self, source: SocketAddr, target: SocketAddr) -> Option<Connection> {
-        self.connections.read().get(&Connection::new(source, target)).copied()
-    }
-
-    /// Returns a snapshot of all the connections.
-    pub fn connections(&self) -> HashSet<Connection> {
-        self.connections.read().clone()
-    }
-
-    /// Returns a snapshot of all the nodes.
-    pub fn nodes(&self) -> HashMap<SocketAddr, u32> {
-        self.nodes.read().clone()
-    }
-}
-
-#[cfg(test)]
-mod test {
-    use super::*;
-
-    #[test]
-    fn connections_partial_eq() {
-        let a = "12.34.56.78:9000".parse().unwrap();
-        let b = "98.76.54.32:1000".parse().unwrap();
-
-        assert_eq!(Connection::new(a, b), Connection::new(b, a));
-        assert_eq!(Connection::new(a, b), Connection::new(a, b));
-    }
-
-    #[test]
-    fn connections_hash() {
-        use std::collections::hash_map::DefaultHasher;
-
-        let a = "11.11.11.11:1000".parse().unwrap();
-        let b = "22.22.22.22:2000".parse().unwrap();
-
-        let mut h1 = DefaultHasher::new();
-        let mut h2 = DefaultHasher::new();
-
-        let k1 = Connection::new(a, b);
-        let k2 = Connection::new(b, a);
-
-        k1.hash(&mut h1);
-        k2.hash(&mut h2);
-
-        // verify k1 == k2 => hash(k1) == hash(k2)
-        assert_eq!(h1.finish(), h2.finish());
-    }
-}
